+++ src/evaluate.h
@@ -3,6 +3,6 @@
 #include <math.h>
 #include "dist_calculation.h"
 
-void recall_and_ratio(float ** &dataset, float ** &querypoints, int data_dimensionality, int ** &queryknn_results, long int ** &gt, int query_size);
+void recall_and_ratio(float ** &dataset, float ** &querypoints, int data_dimensionality, int ** &queryknn_results, int ** &gt, int query_size);
 
 using namespace std;
